restaurant token validation

+++ pages/api/restaurants/index.ts
@@ -1,15 +1,16 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
 import { REQUEST_METHODS } from '#declarations/enums/REST';
+import { baseRestaurant } from 'utils/baseForms';
 
 import { RestaurantsRequestBody, RestaurantsReturnType } from '#firebase/declarations/types';
-import { COLLECTIONS, MESSAGES } from '#firebase/declarations/enums';
+import { COLLECTIONS, COOKIE_NAMES, MESSAGES } from '#firebase/declarations/enums';
 import { RestaurantSchema } from '#firebase/declarations/schemas';
 import { firestore } from '#firebase/initServerApp';
 
 import { handleError } from '#controllers/api/handleError';
 import { objectContainsSameKeys } from '#controllers/api/validation/objectContainsSameKeys';
-import { baseRestaurant } from '#declarations/baseForms';
+import { verifyToken } from '#controllers/api/validation/verifyToken';
 
 export default async (req: NextApiRequest, res: NextApiResponse<RestaurantsReturnType>) => {
 	switch (req.method) {
@@ -28,8 +29,14 @@ export default async (req: NextApiRequest, res: NextApiResponse<RestaurantsRetur
 
 		case REQUEST_METHODS.POST: {
 			try {
+				const token = req.cookies?.[COOKIE_NAMES.TOKEN];
+				if (!token) return res.status(401).json({ message: MESSAGES.UNAUTHORIZED_NO_TOKEN });
+
 				const { uid, data } = req.body as RestaurantsRequestBody;
 				if (!uid || !data) return res.status(400).json({ message: MESSAGES.RESTAURANTS_MANDATORY_FIELDS_ALL });
+
+				const { uid: tokenUid } = await verifyToken(token);
+				if (uid !== tokenUid) return res.status(401).json({ message: MESSAGES.UNAUTHORIZED_TOKEN });
 
 				const { isValid, errorFields } = objectContainsSameKeys<RestaurantSchema>(data, baseRestaurant);
 				if (!isValid) return res.status(400).json({ message: MESSAGES.ERROR, errorFields });
@@ -47,8 +54,14 @@ export default async (req: NextApiRequest, res: NextApiResponse<RestaurantsRetur
 
 		case REQUEST_METHODS.PATCH: {
 			try {
+				const token = req.cookies?.[COOKIE_NAMES.TOKEN];
+				if (!token) return res.status(401).json({ message: MESSAGES.UNAUTHORIZED_NO_TOKEN });
+
 				const { uid, data } = req.body as RestaurantsRequestBody;
 				if (!uid || !data) return res.status(400).json({ message: MESSAGES.RESTAURANTS_MANDATORY_FIELDS_ALL });
+
+				const { uid: tokenUid } = await verifyToken(token);
+				if (uid !== tokenUid) return res.status(401).json({ message: MESSAGES.UNAUTHORIZED_TOKEN });
 
 				const { isValid, errorFields } = objectContainsSameKeys<RestaurantSchema>(data, baseRestaurant);
 				if (!isValid) return res.status(400).json({ message: MESSAGES.ERROR, errorFields });
@@ -67,8 +80,14 @@ export default async (req: NextApiRequest, res: NextApiResponse<RestaurantsRetur
 
 		case REQUEST_METHODS.PUT: {
 			try {
+				const token = req.cookies?.[COOKIE_NAMES.TOKEN];
+				if (!token) return res.status(401).json({ message: MESSAGES.UNAUTHORIZED_NO_TOKEN });
+
 				const { uid, data } = req.body as RestaurantsRequestBody;
 				if (!uid || !data) return res.status(400).json({ message: MESSAGES.RESTAURANTS_MANDATORY_FIELDS_ALL });
+
+				const { uid: tokenUid } = await verifyToken(token);
+				if (uid !== tokenUid) return res.status(401).json({ message: MESSAGES.UNAUTHORIZED_TOKEN });
 
 				const { isValid, errorFields } = objectContainsSameKeys<RestaurantSchema>(data, baseRestaurant);
 				if (!isValid) return res.status(400).json({ message: MESSAGES.ERROR, errorFields });
@@ -87,8 +106,14 @@ export default async (req: NextApiRequest, res: NextApiResponse<RestaurantsRetur
 
 		case REQUEST_METHODS.DELETE: {
 			try {
+				const token = req.cookies?.[COOKIE_NAMES.TOKEN];
+				if (!token) return res.status(401).json({ message: MESSAGES.UNAUTHORIZED_NO_TOKEN });
+
 				const { uid } = req.body as RestaurantsRequestBody;
 				if (!uid) return res.status(400).json({ message: MESSAGES.RESTAURANTS_MANDATORY_FIELDS_ALL });
+
+				const { uid: tokenUid } = await verifyToken(token);
+				if (uid !== tokenUid) return res.status(401).json({ message: MESSAGES.UNAUTHORIZED_TOKEN });
 
 				const restaurantCol = firestore.collection(COLLECTIONS.RESTAURANTS);
 				await restaurantCol.doc(uid).delete();
